backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
Pillow>=10.0,<11.0
gunicorn>=21.2,<22.0
